--- deep_learning/3_api/bartender/src/modules/classifier/__init__.py
@@ -1,3 +1,0 @@
-# __init__.py contents
-
-__all__ = ['Detector']
