+++ deep_learning/3_api/bartender/src/modules/detector/__init__.py
@@ -1,0 +1,3 @@
+# __init__.py contents
+
+__all__ = ['Detector']
